distributed_loader: populate_column_family: ignore directories other than sstable::is_temp_dir
populate_column_family currently lists only regular files. ignoring all directories.
A later patch in this series allows it to list also directories so to cleanup
the temporary sstable directories, yet valid sub-directories, like staging|upload|snapshots,
may still exist and need to be ignored.

Other kinds of handling, like validating recgnized sub-directories and halting on
unrecognized sub-directories are possible, yet out of scope for this patch(set).

+++ distributed_loader.cc
@@ -539,8 +539,8 @@ future<> distributed_loader::populate_column_family(distributed<database>& db, s
             // push future returned by probe_file/rmdir into an array of futures,
             // so that the supplied callback will not block scan_dir() from
             // reading the next entry in the directory.
-            if (de.type && *de.type == directory_entry_type::directory && sstables::sstable::is_temp_dir(de.name)) {
-                if (engine().cpu_id() == 0) {
+            if (de.type && *de.type == directory_entry_type::directory) {
+                if (engine().cpu_id() == 0 && sstables::sstable::is_temp_dir(de.name)) {
                     fs::path dirpath = sstdir / de.name;
                     dblog.info("Found temporary sstable directory: {}, removing", dirpath);
                     futures.push_back(lister::rmdir(dirpath));
